third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFECompositeElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('in') is "F12"
PASS overComposite1.getAttribute('in') is "F1"
PASS successfullyParsed is true

TEST COMPLETE
